app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pymysql
aiomysql
passlib[argon2]

